+++ RestCountry/RestCountry-interactors/src/test/java/com/iliadmastery/country_interactors/Util.kt
@@ -6,4 +6,4 @@ import com.iliadmastery.country_domain.Country
  * Simulate id auto increment (start at zero)
  */
 fun List<Country>.simulateIdAutoIncrement() =
-    this.mapIndexed { index, country -> country.copy(id = index.toLong()) }
+    this.mapIndexed { index, country -> country.copy(id = index) }
